Add sample count summary to run_MAPPED.sh and update publishDir pattern in main.nf
This commit introduces a summary of downloaded and filtered sample counts in run_MAPPED.sh, enhancing user feedback after Step 4. Additionally, the publishDir directive in the DATA_VALIDATION process of main.nf is updated to specify a pattern for the samplesheet.csv file, improving file management and clarity in output organization.