Update run_MAPPED.sh and main.nf to enhance strain parameter handling
This commit modifies run_MAPPED.sh to accept both `--strain` and its alias `-strain`, improving user flexibility. Additionally, main.nf is updated to ensure the `strain` parameter is never null by initializing it to an empty string, which enhances the robustness of the workflow. These changes improve the overall usability and reliability of the pipeline.